# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/project/create-project": {
      "get": {
        "description": "",
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "project_name": {
                  "example": "any"
                },
                "project_description": {
                  "example": "any"
                },
                "project_mission": {
                  "example": "any"
                },
                "project_vision": {
                  "example": "any"
                },
                "project_worthy_of_note": {
                  "example": "any"
                },
                "project_city": {
                  "example": "any"
                },
                "project_state": {
                  "example": "any"
                },
                "project_country": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/project/create-project-roles": {
      "get": {
        "description": "",
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "project_role_name": {
                  "example": "any"
                },
                "project_role_description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/project/create-project-communication": {
      "get": {
        "description": "",
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "project_comm_website": {
                  "example": "any"
                },
                "project_comm_fb": {
                  "example": "any"
                },
                "project_comm_insta": {
                  "example": "any"
                },
                "project_comm_linkedin": {
                  "example": "any"
                },
                "project_comm_collablink": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/project/project-detail/{projectid}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "projectid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/project/edit-project/{projectid}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "projectid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/project/view-application/{project_teamid}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "project_teamid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}